requirements.txt
streamlit==1.31.0
huggingface_hub==0.23.0
openai==1.12.0
fpdf2==2.7.8
pymongo==4.6.1
Pillow==10.2.0
requests==2.31.0
python-dotenv==1.0.1